Create a sandbox containing small subsets of target datasets.

Create a sandbox containing small subsets of target datasets.

build_sandbox(data_id, bbox, range, non.spatial = NULL, report.only = FALSE)

Arguments

data_id
Element or vector of class character.

bbox
Four-element vector with target bounding box, composed of min. x, min. y, max. x and max. y coordinates.

range
Two-element vector of class character with start/end dates (optional, in yyyy-mm-dd format).

non.spatial
Optional argument. How should non-spatial data be treated? Provide one of the following:

a character element with an SQL query

an integer specifying the number of records to extract

report.only
Logical. If TRUE, the function will solely report on storage requirements?

Value

A mirror of the data structure on the cluster, containing a subset of the target datasets.

Details

The function creates a copy of the datasets specified in data_id within the area defined by bbox and the temporal range determined by range. When report.only is set to TRUE, the function will provide a data.frame reporting on the amount of data per individual file that fits to the search criteria. When set to FALSE, the function will inform the user on the amount of data that will be processed and request an output folder to write the data to. Note that the estimates are fairly accurate for gridded and tabular data. For gridded data, the area within a bbox will be consistently covered by pixels, and a data volume is estimated from the amount of cells within it and the data format. For vector data, this is just an estimate that assumes that the target area is consistently covered by all vector entries in the target dataset. this does not consider vector complexity or the actual amount of overlapping vectors.