MAS database builder.

MAS database builder.

collect_metadata(files)

Arguments

files
Vector listing files to evaluate

Value

RDS compiling relevant metadata for every file in the database

Details

Updates, or creates, RDS with metadata for every file in the database. the function will first search an existing file in the database, following the path set in options. If no file exists, a new one will be created. If a file exists, the function will read it, list all files currently in the database, and update the list if the last modification date differs, or if new files were added.

See also